refactor(language-service): move diagnostic methods to utils
